fix bug 1460707: fixed python 3 lint errors
Fixed linting errors via flake8. Modified code so that test pass in both
python 2 & 3. Most modifications for flake errors were done using find
and replace. New imports were added according to pep8 import
conventions. After linting errors were fixed another crontabber test
failed because explicit type checking wanted 'unicode' type. Unicode
type is only in python 2.A string was passed instead.Checked 2/3 cheat
sheet. Cheat sheet suggested to import from builtins library and use
str. But this failed because the str function doesnt provide the casting
behavior needed. So instead I coerced the string to unicode using its
own decode method. Lastly there was an error of types being used.
Buildin library types vs actual native language types. I used the six
library to import string_types which is a compatibility package for both
python 2 & 3.

+++ webapp-django/crashstats/crashstats/tests/test_views.py
@@ -7,6 +7,7 @@ import datetime
 import json
 import random
 import urlparse
+from builtins import int
 from cStringIO import StringIO
 
 import pyquery
@@ -764,7 +765,7 @@ class TestViews(BaseTestViews):
                 for version in ['20.0', '19.0']:
                     for build_type in ['beta', 'release']:
                         response['hits'].append({
-                            'adi_count': long(random.randint(100, 1000)),
+                            'adi_count': int(random.randint(100, 1000)),
                             'build_type': build_type,
                             'date': start_date,
                             'version': version
@@ -1101,7 +1102,7 @@ class TestViews(BaseTestViews):
                 for version in ['18.0b1', '18.0b2', '19.0']:
                     for build_type in ['beta', 'release']:
                         response['hits'].append({
-                            'adi_count': long(random.randint(100, 1000)),
+                            'adi_count': int(random.randint(100, 1000)),
                             'build_type': build_type,
                             'date': start_date,
                             'version': version
